backend/requirements.txt
# Core Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# OpenAI Integration
openai==1.12.0

# Markdown and YAML Processing
python-frontmatter==1.0.0
markdown==3.5.2
pyyaml==6.0.1

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Environment Variables
python-dotenv==1.0.0

# CORS Support
python-jose[cryptography]==3.3.0

# HTTP Client
httpx==0.26.0

# Date and Time
python-dateutil==2.8.2
